Add files via upload
Tired of not having high resolution versions of these in the colors we need. This is not authorized or official. I've had to vectorize what I could find, but it is far from perfect. That said, at the scale we need to use them, I very much doubt it matters.